fix warnings reported by clippy
`cargo clippy --features remote-attestation,stack-guard,virtio-vsock,virtio-serial,vmcall-interrupt`

+++ src/devices/virtio_serial/src/lib.rs
@@ -755,9 +755,7 @@ impl VirtioSerial {
             .borrow_mut()
             .pop_used(&mut g2h, &mut h2g)?;
 
-        *self
-            .receive_queues_prefill
-            .index_mut(queue_idx as usize / 2) -= h2g.len();
+        *self.receive_queues_prefill.index_mut(queue_idx / 2) -= h2g.len();
         self.receive_data(port_id, &h2g, len)?;
         self.fill_port_queue(port_id)?;
